\name{SPMS-package}
\alias{SPMS-package}
\alias{SPMS}
\docType{package}
\title{
Sub-Pathway Miner Software
}
\description{
Use new method to mine sub-pathway.
}
\details{
\tabular{ll}{
Package: \tab SPMS\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2013-05-06\cr
License: \tab GPL-2\cr
}
Open it and download it}
\author{
Xiaomeng Ni 
Maintainer: ni380910405@yahoo.cn
}
\references{
None
}
\keyword{ package }